cql3/type_json: throw when converted double is out of range
before this change, a 64-bit float point number is compared with
an int64_t integer directly, but C++ compiler implicitly promoted
the integer to a nearest float to compare them. because the design
of IEEE 754, a 64-bit float point number cannot always represent
the value of an int64_t integer. when performing conversion,
`std::numeric_limits<int64_t>::max()`, i.e., 9223372036854775807,
is converted the nearest float point number of 9223372036854775808.F.
and we use it as the upper bound of range of valid values convertable
to int64_t. if the number being converted is actually
9223372036854775808.F. we'd consider it in the range. but it is not.

in this change, we compare these two 64-bit numbers in a more
sophiscated way by

1. comparing the two numbers after casting the integer to float
2. comparing the casted integer with the original number if
   the casted number is equal to the float number

this change addresses a corner case which only happens when
serializing a float number passed to `fromJson()` call and
this float number is sitting at the upper bonund of a bigint.

Fixes #13077
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>