sstables: Move SSTable loading to sstable_directory::sort_sstable()
The reason for this change is that we'll want to fully load the
SSTable only at the destination shard.

Later, sort_sstable() will calculate set of owner shards for a
SSTable by only loading scylla metadata file.

If it turns out that the SSTable belongs to current shard, then
we'll fully load the SSTable using the new and fresh
sstable_directory::load_sstable().

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>