sstables: add tiered_storage
this change adds yet another storage backend to SSTables subsystem.
the storage backend is named "tiered_storage", because it provides
us a way to reduce the storage cost and improve the availablity by
persisting data in different types of storage media. currently,
it supports the local filesystem and remote S3 object storage.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>