sstable: do not call into sstable in filesystem_storage::open()
before this change, filesystem_storage::open() reuses
`sstable::make_component_file_writer()` to create the
temporary toc, it will rename the temporary toc to the
real TOC when sealing the sstable.

but this prevents us from reusing filesystem_storage in
yet another storage backend. as the

1. create temporary
2. rename temporary to toc

dance only applies to filesystem_storage. when
filesystem_storage calls into sstable, it uses the generic
backend to create the temporary toc instead of the underlying
filesystem backend. and if we really want to go this way,
we will have to rename the whatever temporary toc component
created by the generic backend to the toc with the seal()
func. but again, this rename op is only implemented in
the filesystem_storage backend. to mirror this operation
in the another backend does not make sense at all.

so in this change, instead of reusing the
`sstable::make_component_file_writer()`, we just inline
its implementation in filesystem_storage to avoid this
problem. this is also an improvement from the design
perspective, as the storage should not call into its
the higher abstraction -- sstable.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>