compaction: s/make_interpose_consumer/make_consumer/
as we don't always interpose consumer in the implementation,
and the way how the implementation decorate the end consumer should
be hidden in the derived class not exposed as part of the function
signature. so let's use a more general name, and leave it to the
derived class to decide whether to interpose the given consumer.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>