build: extract create_rules_for_mode() out
CMake creates the building system for each mode. in other words,
it create a `build.ninja` for Debug, another `build.ninja` for
Release, etc. in order to call into the building systems created
by CMake, we need to

1. split the code which generates shared building rules/targets
   and the code which generates per-mode rules/targets.
2. reuse the code which generates the rules/targets which are
   not generated by the CMake building system

as the first step, let's extract the bits which prepare the
per-mode rules and targets piecemeal.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>