table: log http status code when memtable flush fails
before this change, we use the general purpose
`operator<<(std::ostream&, const std::exception_ptr&)` to print
out the exception thrown when flushing memtable to sstable,
but the inserter operator only pretty print three cases:

- seastar::netsted_exception
- std::system_error
- std::exception

apparently, it is not realistic to exhaust all or even "popular"
exceptions in the inserter operator. it is the caller's
responsibility to handle the exception, and the exact status code
of an unexpected HTTP status is quite interesting when debugging
a failure when talking to an object storage backend. so let's
check for the exception type, and print out the status code if
it is an httpd::base_exception.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>