docs: install docs with --no-root
when running `make setup`, we could have following failure:
```
Installing the current project: scylla (4.3.0)
The current project could not be installed: No file/folder found for package scylla
If you do not want to install the current project use --no-root
```

because docs is not a proper python project named "scylla",
and do not have a directory structure expected by poetry. what we
expect from poetry, is to manage the dependencies for building
the document.

so, in this change, we pass `--no-root` to `poetry install`, this
skips the root package, which does not exist.

see also https://python-poetry.org/docs/cli/#options-2, for more
details on the command line options of poetry.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>