docs: replace outdated references to "--experimental" with "--experimental-features"
The "--experimental" option was removed in commit f6cca741ea. Using this
deprecated option now causes Scylla to fail with the error:

```
error: the argument ('on') for option '--experimental-features' is invalid
```

Replace documentation references to this option with
`--experimental-features` to prevent user confusion and keep docs in
sync with current functionality.

Fixes scylladb/scylladb#22207

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>