script: add a script to rebuild scylla
This script is designed to rebuild the executables for postmortem
debugging.

In general, scylla keeps the release relocatable tarball in the cloud,
so we can always download them on demand. But it could be challenging
for debugging non-release executables, for instance, if scylla leaves
a coredump when it is tested in CI, it would be handy if we can
reproduce it locally and debug the coredump with the executables.

The script takes a version string that uniquely identifies a build,
creates a git worktree at the exact commit, and configures and builds
scylla with the specified build type. The script supports three build
types: debug, release, and dev, defaulting to release.

Example usage:
  ./scripts/rebuild.sh --version "6.3.0~dev-20250114.2eac7a2d616f"

This will parse the version into:
- 6.3.0 (version number)
- dev (build type)
- 20250114 (date)
- 2eac7a2d616f (commit SHA)

And create a worktree at rebuild-2eac7a2d616f (or at a specified path)
with scylla built for debugging.

Fixes scylladb/scylladb#22338

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>